pyproject.toml
[project]
name = "friday-tony-stark-demo"
version = "0.2.0"
description = "Friday — Autonomous AI agent with Telegram, browser, and bash access"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "httpx",
    "python-dotenv",
    "PyYAML>=6.0",
    # Telegram agent
    "python-telegram-bot>=21.0",
    # Browser automation
    "playwright",
    # Web search (no API key required)
    "duckduckgo-search>=6.0",
    # LLM providers
    "openai>=1.0",
    "ollama>=0.3",
    "google-generativeai",
]

[project.optional-dependencies]
# Desktop voice loop (local STT + Coqui TTS — no LiveKit, no cloud API key needed)
voice = [
    "TTS>=0.22.0",          # Coqui TTS (XTTS-v2, tacotron2, etc.)
    "faster-whisper>=1.0",  # Local Whisper STT
    "sounddevice>=0.4",     # Mic capture + speaker playback
    "soundfile>=0.12",      # WAV I/O
]

[project.scripts]
friday = "server:main"
friday_voice = "voice_agent:main"
friday_agent = "agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server.py", "agent.py"]
